+++ thoth-export-server/src/xml/mod.rs
@@ -7,12 +7,11 @@ use xml::writer::{EmitterConfig, EventWriter, XmlEvent};
 
 const ONIX3_NS: &[(&str, &str)] = &[
     ("release", "3.0"),
-    // TODO: what does this refer to? The link is broken.
     ("xmlns", "http://ns.editeur.org/onix/3.0/reference"),
 ];
 
-const ONIX31_NS: &[(&str, &str)] = &[
-    ("release", "3.1"),
+const ONIX312_NS: &[(&str, &str)] = &[
+    ("release", "3.1.2"),
     ("xmlns", "http://ns.editeur.org/onix/3.1/reference"),
 ];
 
@@ -82,8 +81,8 @@ pub(crate) trait XmlElementBlock<T: XmlSpecification> {
     fn xml_element<W: Write>(&self, w: &mut EventWriter<W>) -> ThothResult<()>;
 }
 
-mod onix31_thoth;
-pub(crate) use onix31_thoth::Onix31Thoth;
+mod onix312_thoth;
+pub(crate) use onix312_thoth::Onix312Thoth;
 mod onix3_thoth;
 pub(crate) use onix3_thoth::Onix3Thoth;
 mod onix3_project_muse;